flask-microservices-products/requirements.txt
Flask==1.1.2
Flask-RESTful==0.3.8
flask-mongoengine==0.9.5
python-dotenv==0.14.0